MAJOR CHANGE: Splitting Repository
In order to keep the codebase smaller, and to allow Alexandria to be more reusable, I'm splitting the repository into two separate repositories. This one -- Alexandria -- will remain as the C# library used to provide an API for interacting with AO3 and other fanfiction websites. The other repository -- Bibliothecary -- will be the application/windows service that allows for reporting and finding new fanfics and posting to Tumblr or doing other things.

Hopefully this will be a good change for the two projects, and will make using one or the other (not really possible with Bibliothecary, but now suddenly possible with Alexandria) much easier.